Create static function on course model to add goal and athenaeum references
This function is related to the update action of an athenaeum or goal.
When they are updated it is also possibile to remove or add
the related courses. If this relation is added or removed it must
reflect also in the respective course model that has been removed or added
from athenaeum or goal.